Move encoding and decoding to separate component
